refactor: create chess package
