+++ src/main/java/chess/domain/chess/Board.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.chess;
 
 import java.util.*;
