+++ src/main/java/chess/domain/chess/MoveRule.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.chess;
 
 public interface MoveRule {
