+++ src/main/java/chess/domain/chess/Piece.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.chess;
 
 import chess.domain.moverule.Empty;
 import chess.domain.moverule.King;
