+++ src/main/java/chess/domain/chess/Square.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.chess;
 
 import chess.exception.GameOverException;
