+++ src/test/java/chess/domain/ColumnTest.java
@@ -1,5 +1,6 @@
 package chess.domain;
 
+import chess.domain.chess.Column;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
