+++ src/test/java/chess/domain/GameTest.java
@@ -1,5 +1,6 @@
 package chess.domain;
 
+import chess.domain.chess.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
