+++ src/test/java/chess/domain/moverule/KingTest.java
@@ -1,7 +1,7 @@
 package chess.domain.moverule;
 
-import chess.domain.MoveRule;
-import chess.domain.Position;
+import chess.domain.chess.MoveRule;
+import chess.domain.chess.Position;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
