+++ src/test/java/chess/domain/moverule/RookTest.java
@@ -1,7 +1,7 @@
 package chess.domain.moverule;
 
-import chess.domain.Piece;
-import chess.domain.Position;
+import chess.domain.chess.Piece;
+import chess.domain.chess.Position;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
