+++ src/test/java/chess/service/RoomServiceTest.java
@@ -4,7 +4,7 @@ import chess.config.DataSource;
 import chess.config.DbConnector;
 import chess.config.TableCreator;
 import chess.dao.RoomDao;
-import chess.domain.Piece;
+import chess.domain.chess.Piece;
 import chess.dto.RoomDto;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
