+++ src/test/java/chess/utils/PositionConverterTest.java
@@ -1,6 +1,6 @@
 package chess.utils;
 
-import chess.domain.Position;
+import chess.domain.chess.Position;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
